refactor(code_gen): extract net builtins into builtins/net.rs
